hosted_relay/requirements.txt

eventlet
Flask
python-socketio



